+++ WoWsShipBuilder.UI/ViewModels/ShipVm/MainWindowViewModel.cs
@@ -1,8 +1,9 @@
-﻿using System.Reactive.Linq;
+﻿using System.Linq;
+using System.Reactive.Linq;
 using System.Threading.Tasks;
 using Splat;
 using WoWsShipBuilder.Core;
-using WoWsShipBuilder.Core.BuildCreator;
+using WoWsShipBuilder.Core.Builds;
 using WoWsShipBuilder.Core.Data;
 using WoWsShipBuilder.Core.DataProvider;
 using WoWsShipBuilder.Core.Extensions;
@@ -37,13 +38,22 @@ namespace WoWsShipBuilder.UI.ViewModels.ShipVm
         public async void OpenSaveBuild()
         {
             Logging.Logger.Info("Saving build");
-            var currentBuild = CreateBuild(CurrentBuildName);
-            string shipName = Locator.Current.GetServiceSafe<ILocalizer>().GetGameLocalization(CurrentShipIndex!).Localization;
-            var dialogResult = await BuildCreationInteraction.Handle(new(AppSettingsHelper.Settings, currentBuild, shipName)) ?? BuildCreationResult.Canceled;
+            string shipName = Locator.Current.GetServiceSafe<ILocalizer>().GetGameLocalization(CurrentShipIndex).Localization;
+            var dialogResult = await BuildCreationInteraction.Handle(new(AppSettingsHelper.Settings, shipName, CurrentBuildName)) ?? BuildCreationResult.Canceled;
             if (!dialogResult.Save)
             {
                 return;
             }
+
+            var currentBuild = CreateBuild(dialogResult.BuildName!);
+            var oldBuild = AppData.Builds.FirstOrDefault(existingBuild => existingBuild.BuildName.Equals(currentBuild.BuildName) && existingBuild.ShipIndex.Equals(currentBuild.ShipIndex));
+            if (oldBuild != null)
+            {
+                Logging.Logger.Info("Removing old build with identical name from list of saved builds to replace with new build.");
+                AppData.Builds.Remove(oldBuild);
+            }
+
+            AppData.Builds.Insert(0, currentBuild);
 
             AppSettingsHelper.Settings.IncludeSignalsForImageExport = dialogResult.IncludeSignals;
             CurrentBuildName = currentBuild.BuildName;