+++ WoWsShipBuilder.ViewModels/ShipVm/ConsumableViewModel.cs
@@ -4,7 +4,7 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using System.Threading.Tasks;
 using DynamicData;
-using WoWsShipBuilder.Core.BuildCreator;
+using WoWsShipBuilder.Core.Builds;
 using WoWsShipBuilder.Core.DataProvider;
 using WoWsShipBuilder.Core.Services;
 using WoWsShipBuilder.DataStructures;
@@ -17,7 +17,7 @@ namespace WoWsShipBuilder.ViewModels.ShipVm;
 /// Data for each slot is stored in an individual <see cref="ConsumableSlotViewModel"/>
 /// while this viewmodel is responsible to store these individual instances and provide methods to easily interact with consumable-related data.
 /// </summary>
-public class ConsumableViewModel : ViewModelBase, IBuildStorable
+public class ConsumableViewModel : ViewModelBase, IBuildComponentProvider
 {
     private readonly IAppDataService appDataService;
     private readonly Ship ship;
@@ -91,7 +91,15 @@ public class ConsumableViewModel : ViewModelBase, IBuildStorable
         var modifiers = new List<(string, float)>();
         foreach (int slot in ActivatedSlots)
         {
-            modifiers.AddRange(ConsumableSlots[slot].SelectedConsumable.Modifiers.Select(entry => (entry.Key, entry.Value)));
+            var consumable = ConsumableSlots[slot].SelectedConsumable;
+            if (consumable.Name.Contains("PCY007"))
+            {
+                modifiers.AddRange(consumable.Modifiers.Select(entry => ("speedBoost_" + entry.Key, entry.Value)));
+            }
+            else
+            {
+                modifiers.AddRange(consumable.Modifiers.Select(entry => (entry.Key, entry.Value)));
+            }
         }
 
         return modifiers;