+++ WoWsShipBuilder.Web/wwwroot/scripts/ChartsHelper.js
@@ -55,7 +55,7 @@ function GetColor(index)
  * @description Add multiple datasets to multiple charts at once.
  * @param {Array<string>} chartsId - The charts ID
  * @param {Array<string>} labels - The labels of the dataset to update
- * @param {Array<Array<number>>} datas - The data of each dataset for each chart.
+ * @param {Array<Array<Array<number>>>} datas - The data of each dataset for each chart.
  * @param {Array<number>} indexes - The indexes to associate to a dataset
  */
 export function BatchAddData(chartsId, datas, labels, indexes)
@@ -97,6 +97,39 @@ export function BatchRemoveData(chartsId, labels)
 }
 
 /**
+ * @description Remove multiple datasets from multiple charts at once.
+ * @param {Array<string>} chartsId - The charts ID
+ * @param {Array<string>} indexes - The indexes of the dataset to remove
+ */
+export function BatchRemoveDataFromIndex(chartsId, indexes)
+{
+    chartsId.forEach((chartId) =>{
+        const chart = Chart.getChart(chartId);
+
+        indexes.forEach((index) => {
+            const shipIndex = chart.data.datasets.findIndex(dataset => {
+                return dataset.index === index;
+            })
+            chart.data.datasets.splice(shipIndex, 1);
+        });
+        chart.update();
+    });
+}
+
+/**
+ * @description Remove all datasets from multiple charts
+ * @param {Array<string>} chartsId - The charts ID
+ */
+export function RemoveAllData(chartsId)
+{
+    chartsId.forEach(chartId => {
+        const chart = Chart.getChart(chartId);
+        chart.data.datasets.splice(0, chart.data.datasets.length);
+        chart.update();
+    });
+}
+
+/**
  * @description Update multiple dataset at once for a specific chart.
  * @param {string} chartId - The chart Id
  * @param {Array<string>} labels - The labels of the dataset to update
@@ -112,6 +145,49 @@ export function BatchUpdateData(chartId, labels, newDatas)
         chart.data.datasets[shipIndex].data = newDatas[index];
     });
     chart.update();
+}
+
+/**
+ * @description Update multiple dataset at once for a specific chart.
+ * @param {string} chartId - The chart Id
+ * @param {Array<string>} labels - The labels of the dataset to update
+ * @param {Array<Array<number>>} newDatas - The new data
+ * @param {Array<string>} newLabels - The labels of the dataset to update
+ */
+export function BatchUpdateDataNewLabels(chartId, labels, newDatas, newLabels)
+{
+    const chart = Chart.getChart(chartId);
+    labels.forEach((label, index) => {
+        const shipIndex = chart.data.datasets.findIndex(dataset => {
+            return dataset.label === label;
+        })
+        chart.data.datasets[shipIndex].data = newDatas[index];
+        chart.data.datasets[shipIndex].label = newLabels[index];
+    });
+    chart.update();
+}
+
+/**
+ * @description Update multiple dataset at once for a specific chart.
+ * @param {Array<string>} chartsId - The charts Id to update the data of
+ * @param {Array<number>} indexes - The indexes of the dataset to replace
+ * @param {Array<string>} labels - The new labels
+ * @param {Array<Array<Array<number>>>} newDatas - The new data
+ */
+export function BatchUpdateDataFromIndexes(chartsId, indexes, labels, newDatas)
+{
+    chartsId.forEach((chartId, chartIndex) =>
+    {
+        const chart = Chart.getChart(chartId);
+        indexes.forEach((datasetIndex, index) => {
+            const shipIndex = chart.data.datasets.findIndex(dataset => {
+                return dataset.index === datasetIndex;
+            })
+            chart.data.datasets[shipIndex].data = newDatas[chartIndex][index];
+            chart.data.datasets[shipIndex].label = labels[chartIndex];
+        });
+        chart.update();  
+    });
 }
 
 /**